% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawTable.R
\name{drawTable}
\alias{drawTable}
\title{Draw table of FDSLRM parameters estimates and frequencies}
\usage{
drawTable(
  type,
  periodogram,
  frequencies = NULL,
  fixed_eff,
  random_eff,
  variances
)
}
\arguments{
\item{type}{one of the following values if acceptable: \code{"periodogram", "fixed", "random", "variance"}.}

\item{periodogram}{output of function \code{spec.pgram}.}

\item{frequencies}{vector of (Fourier) frequencies in the form of fractions.}

\item{fixed_eff}{vector of fixed effects.}

\item{random_eff}{vector of random effects.}

\item{variances}{vector of variances. First element should be a variance of white noise
and the next elements should be variances of random effects.}
}
\value{
Table with corresponding values.
}
\description{
\code{drawTable(type, periodogram, frequencies, fixed_eff, random_eff, variances)}
creates a good looking table for estimates of FDSLRM parameters and frequencies.
This function is designed for Jupyter notebooks.
}
\note{
Ver.: 12-Jan-2019 17:20:30.
}
\examples{
# You can see how to use this utility function by opening the following link:
# link to our jupyter notebooks with examples.
}
