% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTS.R
\name{genTS}
\alias{genTS}
\title{Generate FDSLRM time series realization}
\usage{
genTS(mean, V, var)
}
\arguments{
\item{mean}{vector of mean values, \eqn{mean = F\beta}.}

\item{V}{design matrix for random effects.}

\item{var}{vector of variance parameters,
the first one is for the white noise and the next are for the random effects.}
}
\value{
Realization of particular FDSLRM time series.
}
\description{
\code{genTS(mean, V, var)} generates a realization of a time series (FDSLRM) with mean value \code{mean},
variance parameters of white noise and random effects \code{var} and design matrix for random effects \code{V}.
Normal distribution of white noise and random effects is assumed.
}
\details{
explain the form of FDSLRM time series realization as linear mixed model?
}
\note{
Ver.: 03-Jan-2019 18:56:51.
}
\examples{
## EXAMPLE 1
# realization from general FDSLRM
beta <- c (-3, 8, 7)
times <- 1:24
matF <- makeF(times, c(1/24))
matV <- makeV(times, c(1/8, 1/6))
meanv <- matF \%*\% beta
variances <- c(3, 2.7, 1.5, 1, 2)
realization <- genTS(meanv, matV, variances)
print(realization)
plot(times, realization, xlab = "time", type = "o")

## EXAMPLE 2
# realization from classical linear regression time series model (special case of FDSLRM)
beta <- c (-3, 8, 7)
times <- 1:24
matF <- makeF(times, c(1/24))
meanv <- matF \%*\% beta
variances <- 3
realization <- genTS(meanv, var = variances)
print(realization)
plot(times, realization, xlab = "time", type = "o")
}
