% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTS2.R
\name{genTS2}
\alias{genTS2}
\title{Generate FDSLRM time series realization}
\usage{
genTS2(model_formula, beta, nu, n)
}
\arguments{
\item{model_formula}{object of class \code{formula} which contains a symbolic model formula
including functions \eqn{f_i(t)} and \eqn{v_j(t)} neccessary to create design matrices \eqn{F}, \eqn{V}
for fixed and random effects in FDSLRM.}

\item{beta}{vector of mean value parameters.}

\item{nu}{vector of variance parameters,
the first one is for the white noise and the next are for the random effects.}

\item{n}{length of realization.}
}
\value{
Realization of particular FDSLRM time series.
}
\description{
\code{genTS2(model_formula, beta, nu)} generates a realization of a time series (FDSLRM) with mean value \code{mean},
variance parameters of white noise and random effects \code{var} and design matrix for random effects \code{V}.
Normal distribution of white noise and random effects is assumed.
}
\details{
explain the form of FDSLRM time series realization as linear mixed model?
}
\note{
Ver.: 06-Jan-2019 19:15:30.
}
\examples{
## EXAMPLE 1
# realization from general FDSLRM
model_formula <- formula(x ~ cos(2 * pi / 24 * t) + sin(2 * pi / 24 * t) +
                                 cos(2 * pi / 8 * t) + sin(2 * pi / 8 * t) +
                                 cos(2 * pi / 6 * t) + sin(2 * pi / 6 * t))
beta <- c (-3, 8, 7)
nu <- c(3, 2.7, 1.5, 1, 2)
times <- 1:24
n <- length(times)
realization <- genTS2(model_formula, beta, nu, n)
print(realization)
plot(times, realization, xlab = "time", type = "o")

## EXAMPLE 2
# realization from classical linear regression time series model (special case of FDSLRM)
model_formula <- formula(x ~ cos(2 * pi / 24 * t) + sin(2 * pi / 24 * t))
beta <- c (-3, 8, 7)
nu <- 3
times <- 1:24
n <- length(times)
realization <- genTS2(model_formula, beta, nu, n)
print(realization)
plot(times, realization, xlab = "time", type = "o")
}
