% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialFDSLRM.R
\name{initialFDSLRM}
\alias{initialFDSLRM}
\title{Initialize R packages and functions}
\usage{
initialFDSLRM(supress_messages = TRUE)
}
\arguments{
\item{supress_messages}{input parameter of type logical indicating
if messages during packages installation and loading should be hidden or not.
Default is \code{supress_messages = TRUE}.}
}
\description{
\code{initialFDSLRM(supress_messages)} loads, installs all necessary R packages and functions
for FDSLRM fitting and diagnostics.
}
\note{
Ver.: 11-Jan-2019 20:33:50.
}
