% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeF.R
\name{makeF}
\alias{makeF}
\title{Design matrix for fixed effects in FDSLRM}
\usage{
makeF(times, freqs)
}
\arguments{
\item{times}{vector of observation times.}

\item{freqs}{vector of (Fourier) frequencies.}
}
\value{
Design matrix \eqn{F} for fixed effects in FDSLRM.
}
\description{
\code{makeF(times, freqs)} creates a design matrix \eqn{F} for fixed effects (including an intercept)
in FDSLRM (for each frequency \eqn{f} produces two columns of \eqn{F}:
\eqn{sin(2*\pi*f*times}), \eqn{cos(2*\pi*f*times}).
}
\note{
Ver.: 02-Jan-2019 13:28:18.
}
\examples{
## EXAMPLE 1
times <- c(1:24)
frequencies <- c(1/24, 1/8, 1/6)
matF <- makeF(times, frequencies)
print(matF)
}
