% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeM_F.R
\name{makeM_F}
\alias{makeM_F}
\title{Projection matrix \eqn{M_F}}
\usage{
makeM_F(F)
}
\arguments{
\item{F}{matrix.}
}
\value{
Projection matrix \eqn{M_F}.
}
\description{
\code{makeM_F(F)} creates a projection matrix \eqn{M_F}. It is definied by the formula
\eqn{M_F = F((F'F)^(-1))F'} and projects into the space orthogonal to the column space
spanned by the columns of matrix \eqn{F}.
}
\note{
Ver.: 02-Jan-2019 13:49:17.
}
\examples{
## EXAMPLE 1
times <- c(1:24)
frequencies <- c(1/24, 1/8, 1/6)
matF <- makeF(times, frequencies)
matMF <- makeM_F(matF)
print(matMF)
}
