% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeV.R
\name{makeV}
\alias{makeV}
\title{Design matrix for random effects in FDSLRM}
\usage{
makeV(times, freqs)
}
\arguments{
\item{times}{vector of observation times.}

\item{freqs}{vector of (Fourier) frequencies.}
}
\value{
Design matrix \eqn{V} for random effects in FDSLRM.
}
\description{
\code{makeV(times, freqs)} creates a design matrix \eqn{V} for random effects (including an intercept)
in FDSLRM (for each frequency \eqn{f} produces two columns of \eqn{V}:
\eqn{sin(2*\pi*f*times}), \eqn{cos(2*\pi*f*times}).
\code{makeV} is a special case of \code{makeF} except for the first column of \eqn{1} for the intercept.
}
\note{
Ver.: 02-Jan-2019 13:41:24.
}
\examples{
## EXAMPLE 1
times <- c(1:24)
frequencies <- c(1/12, 1/8)
matV <- makeV(times, frequencies)
print(matV)
}
