% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNE.R
\name{MNE}
\alias{MNE}
\title{Modified natural estimates}
\usage{
MNE(X, F, V)
}
\arguments{
\item{X}{time series realization.}

\item{F}{design matrix for fixed effects.}

\item{V}{design matrix for random effects.}
}
\value{
Modified natural estimates of variance parameters \eqn{\sigma^2, \sigma_1^2, ..., \sigma_l^2}.
}
\description{
\code{NE(X, F, V)} calculates modified natural estimates (MNE) of variance parameters in general FDSLRM.
MNE are an unbiased version of NE.
}
\note{
Ver.: 21-Jan-2019 20:48:29.
}
\examples{
## EXAMPLE 1
times <- c(1:40)
freq_fixed <- c(1/24)
freq_rand <- c(1/8, 1/6)
matF <- makeF(times, freq_fixed)
matV <- makeV(times, freq_rand)
X <- rnorm(40, 2, 1.25)
mne <- MNE(X, matF, matV)
print(mne)
}
