% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NE.R
\name{NE}
\alias{NE}
\title{Natural estimates}
\usage{
NE(X, F, V)
}
\arguments{
\item{X}{time series realization.}

\item{F}{design matrix for fixed effects.}

\item{V}{design matrix for random effects.}
}
\value{
Natural estimates of variance parameters \eqn{\sigma^2, \sigma_1^2, ..., \sigma_l^2}.
}
\description{
\code{NE(X, F, V)} calculates natural estimates (NE) of variance parameters in general FDSLRM.
}
\note{
Ver.: 21-Jan-2019 20:40:07.
}
\examples{
## EXAMPLE 1
times <- c(1:40)
freq_fixed <- c(1/24)
freq_rand <- c(1/8, 1/6)
matF <- makeF(times, freq_fixed)
matV <- makeV(times, freq_rand)
X <- rnorm(40, 2, 1.25)
ne <- NE(X, matF, matV)
print(ne)
}
