% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNDOOLSE.R
\name{NNDOOLSE}
\alias{NNDOOLSE}
\title{Non-negative double ordinary least squares estimates}
\usage{
NNDOOLSE(X, F, V)
}
\arguments{
\item{X}{time series realization.}

\item{F}{design matrix for fixed effects.}

\item{V}{design matrix for random effects.}
}
\value{
NNDOOLSE of variance parameters \eqn{\sigma_1^2, ..., \sigma_l^2, \sigma^2}.
}
\description{
\code{NNDOOLSE(X, F, V)} calculates non-negative double ordinary least squares estimates (NNDOOLSE) of variance parameters in FDSLRM.
}
\note{
Ver.: 06-Feb-2019 08:51:47.
}
\examples{
## EXAMPLE 1
times <- c(1:24)
freq_fixed <- c(1/24)
freq_rand <- c(1/8, 1/6)
matF <- makeF(times, freq_fixed)
matV <- makeV(times, freq_rand)
X <- c(40.3, 40.7, 38.5, 37.9, 38.6, 41.1, 45.2, 45.7, 46.7, 46.5, 45.2, 45.1, 45.8, 46.3, 47.5, 48.5, 49.1, 51.7, 50.6, 48, 44.7, 41.2, 40, 40.3)
nndoolse <- NNDOOLSE(X, matF, matV)
print(nndoolse)
}
